i18n-translator/requirements.txt
# i18n-translator dependencies
google-generativeai>=0.3.0

# Optional: for better JSON handling
# jq (install via: brew install jq)
